requirements.txt
Django==2.2.8
Pillow==6.2.1
grab==0.6.41
pyTelegramBotAPI==3.6.6
celery[redis]==4.3.0
mysqlclient==1.4.4
six==1.12.0
textract==1.6.3
djangorestframework==3.11.0
django-cors-headers==3.2.0
django-filter==2.2.0
coreapi==2.3.3
markdown==3.1.1
